Move taxonomy parsing into index creation process
